requirements.txt
# python==3.10
numpy==1.23.4
# pip install torch==2.5.1 torchvision==0.20.1 torchaudio==2.5.1 --index-url https://download.pytorch.org/whl/cu124
torch==2.5.1 --index-url https://download.pytorch.org/whl/cu124
torchvision==0.20.1 --index-url https://download.pytorch.org/whl/cu124
torchaudio==2.5.1 --index-url https://download.pytorch.org/whl/cu124

funasr==1.1.3

uvicorn==0.30.6
fastapi==0.110.1
python-multipart
modelscope
python-dotenv
websockets

## GPT_SoVITS
ffmpeg-python
transformers==4.45.1
pytorch_lightning==2.4.0
py3langid==0.2.2
LangSegment==0.3.5
cn2an==0.5.22
pypinyin==0.53.0
jieba_fast==0.53
onnxruntime==1.17.1; sys_platform == 'darwin'
onnxruntime-gpu==1.19.2; sys_platform != 'darwin'
opencc==1.1.9; sys_platform != 'linux'
opencc==1.1.1; sys_platform == 'linux'

## CosyVoice
conformer==0.3.2
# deepspeed==0.14.2; sys_platform == 'linux'
diffusers==0.29.0
lightning==2.2.4

rich==13.7.1
inflect==7.3.1
hyperpyyaml==1.2.2
openai-whisper==20231117
# conda install -c conda-forge pynini=2.1.5 -y
WeTextProcessing==1.0.3
# python -c "import torch; print(torch.cuda.is_available())"

## LLM
openai==1.61.1
